rap comments are reportable

+++ src/server/api/routers/report.ts
@@ -1,4 +1,5 @@
 import { ReportType, ReportedEntity } from '@prisma/client';
+import { TRPCError } from '@trpc/server';
 import { createTRPCRouter, protectedProcedure } from 'src/server/api/trpc';
 import { z } from 'zod';
 
@@ -10,12 +11,25 @@ export const reportRouter = createTRPCRouter({
         reportedEntity: z.nativeEnum(ReportedEntity),
         rapId: z.string().optional(),
         threadId: z.string().optional(),
-        commentId: z.string().optional(),
+        threadCommentId: z.string().optional(),
         reportedId: z.string().optional()
       })
     )
     .mutation(async ({ ctx, input }) => {
-      const { type, rapId, threadId, commentId, reportedId, reportedEntity } = input;
+      const { type, rapId, threadId, threadCommentId, reportedId, reportedEntity } = input;
+
+      const user = await ctx.prisma.user.findUnique({
+        where: {
+          id: ctx.session.user.id
+        }
+      });
+
+      if (!user?.isAdmin) {
+        throw new TRPCError({
+          code: 'UNAUTHORIZED',
+          message: 'You are not authorized to perform this action'
+        });
+      }
 
       let reportedUserId = reportedId;
 
@@ -28,12 +42,21 @@ export const reportRouter = createTRPCRouter({
         reportedUserId = rap?.userId;
       }
 
+      if (reportedEntity === ReportedEntity.RAP_COMMENT && threadCommentId) {
+        const threadComment = await ctx.prisma.threadComment.findUnique({
+          where: {
+            id: threadCommentId
+          }
+        });
+        reportedUserId = threadComment?.userId;
+      }
+
       const report = await ctx.prisma.report.create({
         data: {
           type,
           rapId,
           threadId,
-          commentId,
+          commentId: threadCommentId,
           reportedId: reportedUserId,
           reporterId: ctx.session.user.id,
           reportedEntity: reportedEntity
@@ -47,11 +70,25 @@ export const reportRouter = createTRPCRouter({
       include: {
         rap: true,
         reporter: true,
-        reported: true
+        reported: true,
+        threadComment: true
       }
     });
   }),
   deleteReport: protectedProcedure.input(z.string()).mutation(async ({ ctx, input }) => {
+    const user = await ctx.prisma.user.findUnique({
+      where: {
+        id: ctx.session.user.id
+      }
+    });
+
+    if (!user?.isAdmin) {
+      throw new TRPCError({
+        code: 'UNAUTHORIZED',
+        message: 'You are not authorized to perform this action'
+      });
+    }
+
     return ctx.prisma.report.delete({
       where: {
         id: input
@@ -59,6 +96,19 @@ export const reportRouter = createTRPCRouter({
     });
   }),
   deleteReports: protectedProcedure.input(z.array(z.string())).mutation(async ({ ctx, input }) => {
+    const user = await ctx.prisma.user.findUnique({
+      where: {
+        id: ctx.session.user.id
+      }
+    });
+
+    if (!user?.isAdmin) {
+      throw new TRPCError({
+        code: 'UNAUTHORIZED',
+        message: 'You are not authorized to perform this action'
+      });
+    }
+
     return ctx.prisma.report.deleteMany({
       where: {
         id: {
